Creates separate file for UIView extension
