+++ CustomDropDown/Extensions/UIViewExtension.swift
@@ -1,0 +1,46 @@
+//
+//  UIViewExtension.swift
+//  CustomDropDown
+//
+//  Created by Jacqueline Alves on 22/10/20.
+//  Copyright © 2020 Amirthy Tejeshwar. All rights reserved.
+//
+
+import UIKit
+
+extension UIView {
+    func addAnchors(top: NSLayoutYAxisAnchor? = nil,
+                    bottom: NSLayoutYAxisAnchor? = nil,
+                    left: NSLayoutXAxisAnchor? = nil,
+                    right: NSLayoutXAxisAnchor? = nil,
+                    padding: CGFloat = 5,
+                    widthConstraint: NSLayoutConstraint? = nil,
+                    heightConstraint: NSLayoutConstraint? = nil) {
+        
+        translatesAutoresizingMaskIntoConstraints = false
+        
+        if let top = top {
+            topAnchor.constraint(equalTo: top, constant: padding).isActive = true
+        }
+
+        if let right = right {
+            rightAnchor.constraint(equalTo: right, constant: -padding).isActive = true
+        }
+        
+        if let left = left {
+            leftAnchor.constraint(equalTo: left, constant: padding).isActive = true
+        }
+        
+        if let bottom = bottom {
+            bottomAnchor.constraint(equalTo: bottom, constant: -padding).isActive = true
+        }
+        
+        if let widthConstraint = widthConstraint {
+            widthConstraint.isActive = true
+        }
+        
+        if let heightConstraint = heightConstraint {
+            heightConstraint.isActive = true
+        }
+    }
+}
